Finish module 5
These are cool, would be nice to review later. I think maybe in a team
setting where other engineers are trying to learn these things, it'd be
useful to watch through one of these at the start of the day (or read
about it, whatever makes more sense for a big group) and then during the
workday share on chat examples of where during the day we saw they could
be used. Maybe make a prize for whoever gets the first "good" example, letting other folks vote on it. Maybe instead of 21 different prizes, give a prize for whoever gets the most first good examples? Just some ideas.